# pkg/api/lookout/api.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "title": "pkg/api/lookout/lookout.proto",
    "version": "version not set"
  },
  "paths": {
    "/api/v1/lookout/cancel_job_set": {
      "post": {
        "tags": [
          "Lookout"
        ],
        "operationId": "CancelJobSet",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lookoutCancelJobSetRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lookoutCancelJobResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/api/v1/lookout/jobs": {
      "post": {
        "tags": [
          "Lookout"
        ],
        "operationId": "GetJobs",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lookoutGetJobsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lookoutGetJobsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/api/v1/lookout/jobsets": {
      "post": {
        "tags": [
          "Lookout"
        ],
        "operationId": "GetJobSets",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lookoutGetJobSetsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lookoutGetJobSetsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/api/v1/lookout/overview": {
      "get": {
        "tags": [
          "Lookout"
        ],
        "operationId": "Overview",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lookoutSystemOverview"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "apiIngressConfig": {
      "type": "object",
      "properties": {
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "certName": {
          "type": "string"
        },
        "ports": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          }
        },
        "tlsEnabled": {
          "type": "boolean"
        },
        "type": {
          "$ref": "#/definitions/apiIngressType"
        },
        "useClusterIP": {
          "type": "boolean"
        }
      }
    },
    "apiIngressType": {
      "description": "Ingress type is being kept here to maintain backwards compatibility for a while.",
      "type": "string",
      "default": "Ingress",
      "enum": [
        "Ingress"
      ]
    },
    "apiJob": {
      "type": "object",
      "properties": {
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "clientId": {
          "type": "string"
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string"
        },
        "ingress": {
          "description": "Services can be provided either as Armada-specific config objects or as proper k8s objects.\nThese options are exclusive, i.e., if either ingress or services is provided,\nthen neither of k8s_ingress or k8s_service can be provided, and vice versa.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiIngressConfig"
          }
        },
        "jobSetId": {
          "type": "string"
        },
        "k8sIngress": {
          "description": "repeated github.com.G-Research.armada.internal.events.KubernetesObject objects = 17;\ngithub.com.G-Research.armada.internal.events.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Ingress"
          }
        },
        "k8sService": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Service"
          }
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "namespace": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "podSpec": {
          "$ref": "#/definitions/v1PodSpec"
        },
        "podSpecs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1PodSpec"
          }
        },
        "priority": {
          "type": "number",
          "format": "double"
        },
        "queue": {
          "type": "string"
        },
        "queueOwnershipUserGroups": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "requiredNodeLabels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiServiceConfig"
          }
        }
      }
    },
    "apiServiceConfig": {
      "type": "object",
      "properties": {
        "ports": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          }
        },
        "type": {
          "$ref": "#/definitions/apiServiceType"
        }
      }
    },
    "apiServiceType": {
      "type": "string",
      "default": "NodePort",
      "enum": [
        "NodePort",
        "Headless"
      ]
    },
...
    "lookoutCancelJobResponse": {
      "type": "object"
    },
    "lookoutCancelJobSetRequest": {
      "type": "object",
      "properties": {
        "jobSet": {
          "type": "string"
        },
        "jobStates": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "queue": {
          "type": "string"
        }
      }
    },
    "lookoutDurationStats": {
      "type": "object",
      "properties": {
        "average": {
          "type": "string"
        },
        "longest": {
          "type": "string"
        },
        "median": {
          "type": "string"
        },
        "q1": {
          "type": "string"
        },
        "q3": {
          "type": "string"
        },
        "shortest": {
          "type": "string"
        }
      }
    },
    "lookoutGetJobSetsRequest": {
      "type": "object",
      "properties": {
        "activeOnly": {
          "type": "boolean"
        },
        "newestFirst": {
          "type": "boolean"
        },
        "queue": {
          "type": "string"
        }
      }
    },
    "lookoutGetJobSetsResponse": {
      "type": "object",
      "properties": {
        "jobSetInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/lookoutJobSetInfo"
          }
        }
      }
    },
    "lookoutGetJobsRequest": {
      "type": "object",
      "properties": {
        "jobId": {
          "type": "string"
        },
        "jobSetIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "jobStates": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "newestFirst": {
          "type": "boolean"
        },
        "owner": {
          "type": "string"
        },
        "queue": {
          "type": "string"
        },
        "skip": {
          "type": "integer",
          "format": "int64"
        },
        "take": {
          "type": "integer",
          "format": "int64"
        },
        "userAnnotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "lookoutGetJobsResponse": {
      "type": "object",
      "properties": {
        "jobInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/lookoutJobInfo"
          }
        }
      }
    },
    "lookoutJobInfo": {
      "type": "object",
      "properties": {
        "cancelled": {
          "type": "string",
          "format": "date-time"
        },
        "job": {
          "$ref": "#/definitions/apiJob"
        },
        "jobJson": {
          "type": "string"
        },
        "jobState": {
          "type": "string"
        },
        "runs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/lookoutRunInfo"
          }
        }
      }
    },
    "lookoutJobSetInfo": {
      "type": "object",
      "properties": {
        "jobSet": {
          "type": "string"
        },
        "jobsFailed": {
          "type": "integer",
          "format": "int64"
        },
        "jobsPending": {
          "type": "integer",
          "format": "int64"
        },
        "jobsQueued": {
          "type": "integer",
          "format": "int64"
        },
        "jobsRunning": {
          "type": "integer",
          "format": "int64"
        },
        "jobsSucceeded": {
          "type": "integer",
          "format": "int64"
        },
        "queue": {
          "type": "string"
        },
        "queuedStats": {
          "$ref": "#/definitions/lookoutDurationStats"
        },
        "runningStats": {
          "$ref": "#/definitions/lookoutDurationStats"
        },
        "submitted": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "lookoutQueueInfo": {
      "type": "object",
      "properties": {
        "jobsPending": {
          "type": "integer",
          "format": "int64"
        },
        "jobsQueued": {
          "type": "integer",
          "format": "int64"
        },
        "jobsRunning": {
          "type": "integer",
          "format": "int64"
        },
        "longestRunningDuration": {
          "type": "string"
        },
        "longestRunningJob": {
          "$ref": "#/definitions/lookoutJobInfo"
        },
        "oldestQueuedDuration": {
          "type": "string"
        },
        "oldestQueuedJob": {
          "$ref": "#/definitions/lookoutJobInfo"
        },
        "queue": {
          "type": "string"
        }
      }
    },
    "lookoutRunInfo": {
      "type": "object",
      "properties": {
        "cluster": {
          "type": "string"
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "error": {
          "type": "string"
        },
        "finished": {
          "type": "string",
          "format": "date-time"
        },
        "k8sId": {
          "type": "string"
        },
        "node": {
          "type": "string"
        },
        "podNumber": {
          "type": "integer",
          "format": "int32"
        },
        "runState": {
          "type": "string"
        },
        "started": {
          "type": "string",
          "format": "date-time"
        },
        "succeeded": {
          "type": "boolean"
        },
        "unableToSchedule": {
          "type": "boolean"
        }
      }
    },
    "lookoutSystemOverview": {
      "type": "object",
      "properties": {
        "queues": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/lookoutQueueInfo"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
...
    "runtimeError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        },
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "v1Ingress": {
      "description": "Ingress is a collection of rules that allow inbound connections to reach the\nendpoints defined by a backend. An Ingress can be configured to give services\nexternally-reachable urls, load balance traffic, terminate SSL, offer name\nbased virtual hosting etc.",
      "type": "object",
      "properties": {
        "metadata": {
          "title": "Standard object's metadata.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata\n+optional",
          "$ref": "#/definitions/v1ObjectMeta"
        },
        "spec": {
          "title": "Spec is the desired state of the Ingress.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional",
          "$ref": "#/definitions/v1IngressSpec"
        },
        "status": {
          "title": "Status is the current state of the Ingress.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional",
          "$ref": "#/definitions/v1IngressStatus"
        }
      }
    },
    "v1IngressBackend": {
      "description": "IngressBackend describes all endpoints for a given service and port.",
      "type": "object",
      "properties": {
        "resource": {
          "title": "Resource is an ObjectRef to another Kubernetes resource in the namespace\nof the Ingress object. If resource is specified, a service.Name and\nservice.Port must not be specified.\nThis is a mutually exclusive setting with \"Service\".\n+optional",
          "$ref": "#/definitions/v1TypedLocalObjectReference"
        },
        "service": {
          "title": "Service references a Service as a Backend.\nThis is a mutually exclusive setting with \"Resource\".\n+optional",
          "$ref": "#/definitions/v1IngressServiceBackend"
        }
      }
    },
...
    "v1IngressSpec": {
      "description": "IngressSpec describes the Ingress the user wishes to exist.",
      "type": "object",
      "properties": {
        "defaultBackend": {
          "title": "DefaultBackend is the backend that should handle requests that don't\nmatch any rule. If Rules are not specified, DefaultBackend must be specified.\nIf DefaultBackend is not set, the handling of requests that do not match any\nof the rules will be up to the Ingress controller.\n+optional",
          "$ref": "#/definitions/v1IngressBackend"
        },
        "ingressClassName": {
          "type": "string",
          "title": "IngressClassName is the name of the IngressClass cluster resource. The\nassociated IngressClass defines which controller will implement the\nresource. This replaces the deprecated `kubernetes.io/ingress.class`\nannotation. For backwards compatibility, when that annotation is set, it\nmust be given precedence over this field. The controller may emit a\nwarning if the field and annotation have different values.\nImplementations of this API should ignore Ingresses without a class\nspecified. An IngressClass resource may be marked as default, which can\nbe used to set a default value for this field. For more information,\nrefer to the IngressClass documentation.\n+optional"
        },
        "rules": {
          "type": "array",
          "title": "A list of host rules used to configure the Ingress. If unspecified, or\nno rule matches, all traffic is sent to the default backend.\n+listType=atomic\n+optional",
          "items": {
            "$ref": "#/definitions/v1IngressRule"
          }
        },
        "tls": {
          "type": "array",
          "title": "TLS configuration. Currently the Ingress only supports a single TLS\nport, 443. If multiple members of this list specify different hosts, they\nwill be multiplexed on the same port according to the hostname specified\nthrough the SNI TLS extension, if the ingress controller fulfilling the\ningress supports SNI.\n+listType=atomic\n+optional",
          "items": {
            "$ref": "#/definitions/v1IngressTLS"
          }
        }
      }
    },
    "v1IngressStatus": {
      "description": "IngressStatus describe the current state of the Ingress.",
      "type": "object",
      "properties": {
        "loadBalancer": {
          "title": "LoadBalancer contains the current status of the load-balancer.\n+optional",
          "$ref": "#/definitions/v1LoadBalancerStatus"
        }
      }
    },
...
    "v1Service": {
      "description": "Service is a named abstraction of software service (for example, mysql) consisting of local port\n(for example 3306) that the proxy listens on, and the selector that determines which pods\nwill answer requests sent through the proxy.",
      "type": "object",
      "properties": {
        "metadata": {
          "title": "Standard object's metadata.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata\n+optional",
          "$ref": "#/definitions/v1ObjectMeta"
        },
        "spec": {
          "title": "Spec defines the behavior of a service.\nhttps://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional",
          "$ref": "#/definitions/v1ServiceSpec"
        },
        "status": {
          "title": "Most recently observed status of the service.\nPopulated by the system.\nRead-only.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional",
          "$ref": "#/definitions/v1ServiceStatus"
        }
      }
    },
...
    "v1ServiceStatus": {
      "description": "ServiceStatus represents the current status of a service.",
      "type": "object",
      "properties": {
        "conditions": {
          "type": "array",
          "title": "Current service state\n+optional\n+patchMergeKey=type\n+patchStrategy=merge\n+listType=map\n+listMapKey=type",
          "items": {
            "$ref": "#/definitions/v1Condition"
          }
        },
        "loadBalancer": {
          "title": "LoadBalancer contains the current status of the load-balancer,\nif one is present.\n+optional",
          "$ref": "#/definitions/v1LoadBalancerStatus"
        }
      }
    },
...
  }
}